% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexistence_curves.r
\name{coexistence_curves}
\alias{coexistence_curves}
\title{Coexistence and species-area curves}
\usage{
coexistence_curves(Y, xy, sample_ids, dirs, nsim = 100)
}
\arguments{
\item{Y}{Virus community data}

\item{xy}{Matrix or data.frame of x and y coordinates}

\item{sample_ids}{Vector of row identifiers that matches with Y and xy}

\item{dirs}{Directories}

\item{nsim}{Number of simulations}
}
\value{
List: 1) (True) coexistences always within the same plant, 2) pooled coexistences, 3) cumulative species richness
}
\description{
Calculate coexistence and species-area curves with different starting points
}
