% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinfs_by_pop.r
\name{coinfs_by_pop}
\alias{coinfs_by_pop}
\title{(Co)infections by population}
\usage{
coinfs_by_pop(Y, pop_lifes, POPs, dirs)
}
\arguments{
\item{Y}{Virus community data}

\item{pop_lifes}{List of selected populations' names, e.g. list("1", "2", ...)}

\item{POPs}{Numerical vector indicating the host populations for Y}

\item{dirs}{Directories, created with 'set_dirs'}
}
\value{
List of results for the analysis
}
\description{
Calculate coinfections for host populations
}
