% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_beta_pair.r
\name{do_beta_pair}
\alias{do_beta_pair}
\title{Pairwise beta diversity for populations}
\usage{
do_beta_pair(Y, POPs, dirs)
}
\arguments{
\item{Y}{Virus community data}

\item{POPs}{Numerical vector indicating the host populations for Y}
}
\value{
List of results for the analysis
}
\description{
Calculate beta diversity for virus populations using the package 'betapart' and function 'beta.pair'
}
