% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_cooccur.r
\name{do_cooccur}
\alias{do_cooccur}
\title{Co-occurrence analysis}
\usage{
do_cooccur(Y, POPs, dirs)
}
\arguments{
\item{Y}{Virus community data}

\item{POPs}{Numerical vector indicating the host populations for Y}
}
\value{
List of results for the analysis
}
\description{
Do co-occurrence analysis using the package and function 'cooccur'
}
