% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pops.r
\name{plot_pops}
\alias{plot_pops}
\title{Plot infection profiles}
\usage{
plot_pops(dat, dirs)
}
\arguments{
\item{dat}{Data, produced by "process_data"}

\item{dirs}{Directories}
}
\value{
Piecharts saved as .png under "figs/infections_by_pop" and "figs/coinfections_by_pop". These folders are created if they do not exist already.
}
\description{
Plot infection and coinfection prefiles of all populations as piecharts
}
