% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.r
\name{process_data}
\alias{process_data}
\title{Process meta17 data}
\usage{
process_data(dirs, return_data = TRUE, save_data = TRUE, rmNAs = TRUE)
}
\arguments{
\item{return_data}{Return the data}

\item{save_data}{Save the data}

\item{rmNAs}{Remove rows containing NAs from the merged data frame}
}
\description{
Modify meta17 data
}
