% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dirs.r
\name{set_dirs}
\alias{set_dirs}
\title{Defining directories}
\usage{
set_dirs(working_dir, fit_fold = "fits")
}
\arguments{
\item{working_dir}{Full path to the working directory, under which everything else will be created.}

\item{fit_fold}{Folder where the model fits will be saved  (defaults to "fits")}
}
\value{
List of directories needed for the modelling pipeline, automatically created as subdirectories under the working_dir
}
\description{
Defining directories/paths to be used
}
